Rename appropriate files for the name switch
The new name for VersionControl is 'hero', as in 'Herodotus', the historian, as I agreed upon with Ken Ellorando
