+++ VersionControl/VersionControl.vcxproj
@@ -23,6 +23,7 @@
     <Keyword>Win32Proj</Keyword>
     <RootNamespace>VersionControl</RootNamespace>
     <WindowsTargetPlatformVersion>10.0.10586.0</WindowsTargetPlatformVersion>
+    <ProjectName>Hero</ProjectName>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'" Label="Configuration">
@@ -180,12 +181,12 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="VersionControl.cpp" />
+    <ClCompile Include="hero.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="crossplatform.h" />
     <ClInclude Include="Utils.h" />
-    <ClInclude Include="VersionControl.h" />
+    <ClInclude Include="hero.h" />
   </ItemGroup>
   <ItemGroup>
     <Text Include="..\commit-blob.txt" />
